requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==9.0.1
cachecontrol==0.12.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
cachetools==4.2.2; python_version ~= '3.5'
cerberus==1.3.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3; python_version >= '3'
click==8.0.1; python_version >= '3.6'
firebase-admin==5.0.1
flask-restful==0.3.9
flask==2.0.1
google-api-core[grpc]==1.31.0; platform_python_implementation != 'PyPy'
google-api-python-client==2.15.0; python_version >= '3.6'
google-auth-httplib2==0.1.0
google-auth==1.33.1; python_version >= '3.6'
google-cloud-core==1.7.1; python_version >= '3.6'
google-cloud-firestore==2.2.0; platform_python_implementation != 'PyPy'
google-cloud-storage==1.41.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-crc32c==1.1.2; python_version >= '3.5'
google-resumable-media==1.3.1; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2; python_version >= '3'
importlib-metadata==4.6.1; python_version < '3.8'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
msgpack==1.0.2
packaging==21.0; python_version >= '3.6'
proto-plus==1.19.0; python_version >= '3.6'
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0; python_version < '3.8'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
zipp==3.5.0; python_version >= '3.6'
